+++ autodonate/lib/templatetags/__init__.py
@@ -1,0 +1,1 @@
+"""Project template tags."""