+++ index/admin.py
@@ -1,0 +1,5 @@
+"""Admin module for index app."""
+
+from django.contrib import admin
+
+# Register your models here.