+++ index/apps.py
@@ -1,0 +1,10 @@
+"""App module for index app."""
+
+from django.apps import AppConfig
+
+
+class IndexConfig(AppConfig):
+    """Index app config."""
+
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "index"